Fixed bug from prev. commit and fixed bug discovered in Character parse
Added the missing romance and charisma stats for Tkrn, Jerci and Zevara.
When a wrong number of primary stats was parsed, no parsing exception
was thrown, instead it got passed to the constructor and an
IllegalArgumentException occurred, so the application crashed without
any user notification. Now it throws the parsing exception.