added and improved documentation, improved exception handling, balancing
Character and CharacterBase are now fully documented except for hashCode
and equals. Changed the IAE to ISE in Registry#add(T) and changed calls
to it accordingly, also prevented these to be thrown in parsing player
things, parse exceptions are thrown instead. Created new ParseException
for trying to add an existing ID.