Fixed a bug that occured with same characters in own and enemy's team
Thought I fixed this bug already, but I was wrong! Just hid it better.
Now it should be fixed for good. That means that the party encounter is
deeply cloned before thrown into battle, and the characters one won
against are healed up after being pseudo copy constructed. That means at
any point the battle contains (pseudo) clones of the player's companions
but never references to them, and these clones are discarded after the
battle is discarded.